refactor: move files into logical folders
