+++ lib/dispatcher/agent.js
@@ -1,12 +1,12 @@
 'use strict'
 
-const { InvalidArgumentError } = require('./core/errors')
-const { kClients, kRunning, kClose, kDestroy, kDispatch, kInterceptors } = require('./core/symbols')
+const { InvalidArgumentError } = require('../core/errors')
+const { kClients, kRunning, kClose, kDestroy, kDispatch, kInterceptors } = require('../core/symbols')
 const DispatcherBase = require('./dispatcher-base')
 const Pool = require('./pool')
 const Client = require('./client')
-const util = require('./core/util')
-const createRedirectInterceptor = require('./interceptor/redirectInterceptor')
+const util = require('../core/util')
+const createRedirectInterceptor = require('../interceptor/redirectInterceptor')
 
 const kOnConnect = Symbol('onConnect')
 const kOnDisconnect = Symbol('onDisconnect')
