+++ lib/dispatcher/balanced-pool.js
@@ -3,7 +3,7 @@
 const {
   BalancedPoolMissingUpstreamError,
   InvalidArgumentError
-} = require('./core/errors')
+} = require('../core/errors')
 const {
   PoolBase,
   kClients,
@@ -13,8 +13,8 @@ const {
   kGetDispatcher
 } = require('./pool-base')
 const Pool = require('./pool')
-const { kUrl, kInterceptors } = require('./core/symbols')
-const { parseOrigin } = require('./core/util')
+const { kUrl, kInterceptors } = require('../core/symbols')
+const { parseOrigin } = require('../core/util')
 const kFactory = Symbol('factory')
 
 const kOptions = Symbol('options')
