+++ lib/dispatcher/client.js
@@ -8,10 +8,10 @@ const assert = require('node:assert')
 const net = require('node:net')
 const http = require('node:http')
 const { pipeline } = require('node:stream')
-const util = require('./core/util')
-const { channels } = require('./core/diagnostics')
-const timers = require('./timers')
-const Request = require('./core/request')
+const util = require('../core/util.js')
+const { channels } = require('../core/diagnostics.js')
+const timers = require('../util/timers.js')
+const Request = require('../core/request.js')
 const DispatcherBase = require('./dispatcher-base')
 const {
   RequestContentLengthMismatchError,
@@ -26,8 +26,8 @@ const {
   HTTPParserError,
   ResponseExceededMaxSizeError,
   ClientDestroyedError
-} = require('./core/errors')
-const buildConnector = require('./core/connect')
+} = require('../core/errors.js')
+const buildConnector = require('../core/connect.js')
 const {
   kUrl,
   kReset,
@@ -79,7 +79,7 @@ const {
   kHTTP2BuildRequest,
   kHTTP2CopyHeaders,
   kHTTP1BuildRequest
-} = require('./core/symbols')
+} = require('../core/symbols.js')
 
 /** @type {import('http2')} */
 let http2
@@ -112,13 +112,13 @@ const FastBuffer = Buffer[Symbol.species]
 const kClosedResolve = Symbol('kClosedResolve')
 
 /**
- * @type {import('../types/client').default}
+ * @type {import('../../types/client.js').default}
  */
 class Client extends DispatcherBase {
   /**
    *
    * @param {string|URL} url
-   * @param {import('../types/client').Client.Options} options
+   * @param {import('../../types/client.js').Client.Options} options
    */
   constructor (url, {
     interceptors,
@@ -472,16 +472,16 @@ function onHTTP2GoAway (code) {
   resume(client)
 }
 
-const constants = require('./llhttp/constants')
-const createRedirectInterceptor = require('./interceptor/redirectInterceptor')
+const constants = require('../llhttp/constants.js')
+const createRedirectInterceptor = require('../interceptor/redirectInterceptor.js')
 const EMPTY_BUF = Buffer.alloc(0)
 
 async function lazyllhttp () {
-  const llhttpWasmData = process.env.JEST_WORKER_ID ? require('./llhttp/llhttp-wasm.js') : undefined
+  const llhttpWasmData = process.env.JEST_WORKER_ID ? require('../llhttp/llhttp-wasm.js') : undefined
 
   let mod
   try {
-    mod = await WebAssembly.compile(require('./llhttp/llhttp_simd-wasm.js'))
+    mod = await WebAssembly.compile(require('../llhttp/llhttp_simd-wasm.js'))
   } catch (e) {
     /* istanbul ignore next */
 
@@ -489,7 +489,7 @@ async function lazyllhttp () {
     // being enabled, but the occurring of this other error
     // * https://github.com/emscripten-core/emscripten/issues/11495
     // got me to remove that check to avoid breaking Node 12.
-    mod = await WebAssembly.compile(llhttpWasmData || require('./llhttp/llhttp-wasm.js'))
+    mod = await WebAssembly.compile(llhttpWasmData || require('../llhttp/llhttp-wasm.js'))
   }
 
   return await WebAssembly.instantiate(mod, {
@@ -1500,7 +1500,7 @@ function write (client, request) {
 
   if (util.isFormDataLike(body)) {
     if (!extractBody) {
-      extractBody = require('./web/fetch/body.js').extractBody
+      extractBody = require('../web/fetch/body.js').extractBody
     }
 
     const [bodyStream, contentType] = extractBody(body)
