+++ lib/dispatcher/dispatcher-base.js
@@ -5,8 +5,8 @@ const {
   ClientDestroyedError,
   ClientClosedError,
   InvalidArgumentError
-} = require('./core/errors')
-const { kDestroy, kClose, kDispatch, kInterceptors } = require('./core/symbols')
+} = require('../core/errors')
+const { kDestroy, kClose, kDispatch, kInterceptors } = require('../core/symbols')
 
 const kDestroyed = Symbol('destroyed')
 const kClosed = Symbol('closed')
