+++ lib/dispatcher/pool-stats.js
@@ -1,4 +1,4 @@
-const { kFree, kConnected, kPending, kQueued, kRunning, kSize } = require('./core/symbols')
+const { kFree, kConnected, kPending, kQueued, kRunning, kSize } = require('../core/symbols')
 const kPool = Symbol('pool')
 
 class PoolStats {
