+++ lib/dispatcher/pool.js
@@ -10,10 +10,10 @@ const {
 const Client = require('./client')
 const {
   InvalidArgumentError
-} = require('./core/errors')
-const util = require('./core/util')
-const { kUrl, kInterceptors } = require('./core/symbols')
-const buildConnector = require('./core/connect')
+} = require('../core/errors')
+const util = require('../core/util')
+const { kUrl, kInterceptors } = require('../core/symbols')
+const buildConnector = require('../core/connect')
 
 const kOptions = Symbol('options')
 const kConnections = Symbol('connections')
