+++ lib/dispatcher/retry-agent.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Dispatcher = require('./dispatcher')
-const RetryHandler = require('./handler/RetryHandler')
+const RetryHandler = require('../handler/RetryHandler')
 
 class RetryAgent extends Dispatcher {
   #agent = null
