+++ lib/util/timers.js
