+++ test/client-keep-alive.js
@@ -4,7 +4,7 @@ const { tspl } = require('@matteo.collina/tspl')
 const { test, after } = require('node:test')
 const { once } = require('node:events')
 const { Client } = require('..')
-const timers = require('../lib/timers')
+const timers = require('../lib/util/timers')
 const { kConnect } = require('../lib/core/symbols')
 const { createServer } = require('node:net')
 const http = require('node:http')
