+++ test/proxy-agent.js
@@ -6,8 +6,8 @@ const { request, fetch, setGlobalDispatcher, getGlobalDispatcher } = require('..
 const { InvalidArgumentError } = require('../lib/core/errors')
 const { readFileSync } = require('node:fs')
 const { join } = require('node:path')
-const ProxyAgent = require('../lib/proxy-agent')
-const Pool = require('../lib/pool')
+const ProxyAgent = require('../lib/dispatcher/proxy-agent')
+const Pool = require('../lib/dispatcher/pool')
 const { createServer } = require('node:http')
 const https = require('node:https')
 const proxy = require('proxy')
